+++ src/lib/utils/assert.js
@@ -1,5 +1,5 @@
 export default function assert(condition, message) {
   if (!condition) {
-    throw new Error(`Assertion failed. ${message}`);
+    throw new Error(message || 'Assertion failed');
   }
 }