+++ test/src/lib/log.spec.js
@@ -117,10 +117,39 @@ test('Log#error', t => {
   t.end();
 });
 
+test('Log#assert', t => {
+  const log = new Log({id: 'test'});
+  t.ok(log instanceof Log, 'log created successfully');
+  t.doesNotThrow(() => log.assert(true, 'test'), 'log.assert works');
+  t.throws(() => log.assert(false, 'test'), 'log.assert works');
+  t.end();
+});
+
 test('Log#table', t => {
   const log = new Log({id: 'test'});
   t.ok(log instanceof Log, 'log created successfully');
   t.doesNotThrow(() => log.table(0, {a: {c: 1}, b: {c: 2}})(), 'log.table works');
   t.doesNotThrow(() => log.table(0, {a: {c: 1}, b: {c: 2}})(), 'log.table(columns) works');
+  t.end();
+});
+
+test('Log#get', t => {
+  const log = new Log({id: 'test'});
+  t.ok(log instanceof Log, 'log created successfully');
+  t.doesNotThrow(() => log.get('priority'), 'log.get(\'priority\') works');
+  t.end();
+});
+
+test('Log#set', t => {
+  const log = new Log({id: 'test'});
+  t.ok(log instanceof Log, 'log created successfully');
+  t.doesNotThrow(() => log.set('priority', 1), 'log.set(\'priority\', 1) works');
+  t.end();
+});
+
+test('Log#status', t => {
+  const log = new Log({id: 'test'});
+  t.ok(log instanceof Log, 'log created successfully');
+  t.doesNotThrow(() => log.settings(), 'log.status() works');
   t.end();
 });