+++ docs/whats-new.md
@@ -1,3 +1,13 @@
-# Version 1
+# What's New
+
+## v2.1
+
+### Log class additions
+- New method `log.assert()`
+- New method `log.settings()`
+- New method `log.set()`
+- New method `log.get()`
+
+## v1.0
 
 probe.gl is now open source!